chore(libs): restructure stripe lib into responsibility-driven structure
