+++ libs/payments/stripe/src/index.ts
@@ -36,7 +36,12 @@ export * from './lib/stripe.config';
 export * from './lib/stripe.constants';
 export * from './lib/stripe.error';
 export * from './lib/stripe.factories';
-export * from './lib/stripe.manager';
+export * from './lib/customer.manager';
+export * from './lib/invoice.manager';
+export * from './lib/price.manager';
+export * from './lib/product.manager';
+export * from './lib/promotionCode.manager';
+export * from './lib/subscription.manager';
 export * from './lib/stripe.service';
 export * from './lib/stripe.types';
 export * from './lib/stripe.util';
